requirements.txt
swarms
swarm-models
swarms-memory
pydantic
loguru
python-dotenv
clusterops
griptape
rich>=13.7.0